+++ src/main/java/com/invadermonky/justenoughmagiculture/JustEnoughMagiculture.java
@@ -22,7 +22,7 @@ public class JustEnoughMagiculture {
     public static final String MOD_ID = "justenoughmagiculture";
     public static final String MOD_ALIAS = "jem";
     public static final String MOD_NAME = "Just Enough Magiculture";
-    public static final String MOD_VERSION = "1.12.2-1.1.5";
+    public static final String MOD_VERSION = "1.12.2-1.2.0";
     public static final String MC_VERSION = "[1.12.2]";
     public static final String DEPENDENCIES = (
             "required-after:" + ConstantIds.JUSTENOUGHITEMS + "@[4.7.0,)" +
@@ -30,7 +30,7 @@ public class JustEnoughMagiculture {
                     ";after:" + ConstantIds.ANIMANIA +
                     ";after:" + ConstantIds.ATUM +
                     ";after:" + ConstantIds.BEAR_WITH_ME +
-                    ";after:" + ConstantIds.BEAST_SLAYER_BETA + "@" + ConstantVersions.BEAST_SLAYER_BETA +
+                    ";after:" + ConstantIds.BEAST_SLAYER + "@" + ConstantVersions.BEAST_SLAYER +
                     ";after:" + ConstantIds.BETWEENLANDS +
                     ";after:" + ConstantIds.BEWITCHMENT +
                     ";after:" + ConstantIds.BOTANIA +
@@ -66,7 +66,11 @@ public class JustEnoughMagiculture {
                     ";after:" + ConstantIds.WADDLES +
 
                     //Mods JEM has to load after because they eat glue.
-                    ";after:" + ConstantIds.LIVINGENCHANTMENT
+                    ";after:" + ConstantIds.ELEMENTAL_ITEMS +
+                    ";after:" + ConstantIds.LIVINGENCHANTMENT +
+                    ";after:" + ConstantIds.LOOTTABLETWEAKER +
+                    ";after:" + ConstantIds.LOOTTWEAKER +
+                    ";after:" + ConstantIds.MYSTCRAFT
     );
 
     public static final String ProxyClientClass = "com.invadermonky.justenoughmagiculture.proxy.ClientProxy";