+++ src/main/java/com/invadermonky/justenoughmagiculture/configs/mods/JEMConfigAstralSorcery.java
@@ -1,0 +1,15 @@
+package com.invadermonky.justenoughmagiculture.configs.mods;
+
+import com.invadermonky.justenoughmagiculture.JustEnoughMagiculture;
+import com.invadermonky.justenoughmagiculture.util.ModIds.ConstantNames;
+import net.minecraftforge.common.config.Config.Comment;
+import net.minecraftforge.common.config.Config.RequiresMcRestart;
+
+public class JEMConfigAstralSorcery {
+    private static final String MOD_NAME = ConstantNames.ASTRAL_SORCERY;
+    private static final String LANG_KEY = "config." + JustEnoughMagiculture.MOD_ALIAS + ":";
+
+    @RequiresMcRestart
+    @Comment("Enables JEM " + MOD_NAME + " JER dungeon integration.")
+    public boolean enableJERDungeons = true;
+}