+++ changelog.md
@@ -2,19 +2,31 @@
 
 ## Changelog
 
+### 1.12.2-1.2.0
+- Added Astral Sorcery integration
+- Added new Oceanic Expanse mobs (version 1.1.x)
+- Fixed missing dungeon chest translation keys
+- Fixed Atum Scarab displaying Camel drops
+- Fixed crash with new releases of Beast Slayer
+<br><br>
+
 ### 1.12.2-1.1.5
 - Fixed a crash with Mystcraft
+  <br><br>
 
 ### 1.12.2-1.1.4
 - Added tallow drop to pigs when Quark Tallow and Candles feature is enabled
 - Modified some code to support Beast Slayer 1.9 and  2.0 when it is released
-  - If using the alpha release of Beast Slayer, you can disable JER integration in the JEM config to fix the crash. 
+  - If using the alpha release of Beast Slayer, you can disable JER integration in the JEM config to fix the crash.
+    <br><br>
 
 ### 1.12.2-1.1.3
 - Added SeriousCreeper's Loot Chest integration to the JEM Loot Bags JEI category
+  <br><br>
 
 ### 1.12.2-1.1.2
 - Fixed a CTD bug caused by LootTableTweaker + LivingEnchantments + JEM interaction
+  <br><br>
 
 ### 1.12.2-1.1.1
 - Added Grimoire of Gaia loot boxes to the JEM Loot Bags JEI category
@@ -23,6 +35,7 @@
   them in the JEM config, or by using Crafttweaker JEI integration.
 - Modified Loot Bags Category to better indicate the lootable item
 - Fixed a rare item bug in the Loot Bags category
+  <br><br>
 
 ### 1.12.2-1.1.0
 
@@ -40,6 +53,7 @@
 - Added Mutant Beasts JER Integration
 - Added Rats JER Integration
 - Added Special Mobs JER Integration
+  <br><br>
 
 ### 1.12.2-1.0.0